add configure method and refactor run
