+++ lib/commands/configure/make-populate-config.js
