+++ lib/commands/configure/make-populate-secret.js
