+++ lib/commands/configure/make-prompt-required-configs.js
