+++ lib/commands/run/make-run.js
@@ -1,28 +1,3 @@
-const { makePopulateConfig } = require('./make-populate-config');
-const { makePopulateSecret } = require('./make-populate-secret');
-const { makePromptRequiredConfigs } = require('./make-prompt-required-configs');
+const makeRun = ({ init, configure }) => () => init().then(configure);
 
-const makeRun = ({ stage, interactive, settingsService, parameterStore }) => () => {
-  const promptRequiredConfigs = makePromptRequiredConfigs({ parameterStore });
-  const populateConfig = makePopulateConfig({ parameterStore, promptRequiredConfigs });
-  const populateSecret = makePopulateSecret({ parameterStore, promptRequiredConfigs});
-
-  return settingsService.getSettings()
-  .tap((settings) => {
-    return populateConfig({
-      settings,
-      stage,
-      interactive
-    });
-  })
-  .then(settings => {
-    return populateSecret({
-      settings,
-      interactive
-    });
-  });
-};
-
-module.exports = {
-  makeRun
-};
+module.exports = { makeRun };
