+++ lib/make-oprah.js
@@ -7,8 +7,9 @@ const { makeParameterStore } = require('./services/parameter-store/make-paramete
 const { makeSettingsService } = require('./services/settings/make-settings-service');
 const { makeCfService } = require('./services//cf/make-cf-service');
 const { makeList } = require('./commands/list/make-list');
-const { makeInit } = require('./commands/init/make-init');
 const { makeRun } = require('./commands/run/make-run');
+const { makeInit } = require('./commands/init/make-init');
+const { makeConfigure } = require('./commands/configure/make-configure');
 
 const makeOprah = ({ stage, config, variables = {}, interactive = false }) => {
   const cfService = makeCfService();
@@ -28,19 +29,24 @@ const makeOprah = ({ stage, config, variables = {}, interactive = false }) => {
   log(chalk.gray(`Stage  --> ${stage}`));
   log(chalk.gray(`Config --> ${config}`));
 
-  return {
-    run: () => {
-      log(chalk.cyan(`Running Oprah in ${interactive ? 'interactive' : 'non-interactive'} mode..`));
+  const configure = () => {
+    log(chalk.cyan(`Running Oprah in ${interactive ? 'interactive' : 'non-interactive'} mode..`));
 
-      return makeRun({
-        interactive,
-        stage,
-        settingsService,
-        parameterStore
-      })();
-    },
+    return makeConfigure({
+      interactive,
+      stage,
+      settingsService,
+      parameterStore
+    })();
+  };
+
+  const init = makeInit({ settingsService, cfService });
+
+  return {
+    configure,
+    init,
+    run: makeRun({ init, configure }),
     list: makeList({ settingsService, parameterStore }),
-    init: makeInit({ settingsService, cfService })
   };
 };
